Load nextPetition in Thanks container
